Potential fix for code scanning alert no. 4: Implicit narrowing conversion in compound assignment

+++ backend/src/main/java/com/aaron212/onlinelibrarymanagement/backend/service/StatisticsService.java
@@ -66,7 +66,7 @@ public class StatisticsService {
         for (IndexCategory category : categories) {
             List<Book> booksInCategory = bookRepository.findByIndexCategory(category);
             int totalCount = 0;
-            int availableCount = 0;
+            long availableCount = 0;
 
             for (Book book : booksInCategory) {
                 List<BookCopy> copies = bookCopyRepository.findByBook(book);
